% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{BoxCoxLambda}
\alias{BoxCoxLambda}
\title{Performs Box-Cox transformation of the data. The transformation parameter (lambda) is chosen to minimize deviation from the normal distribution (minumum sum of squared skewness and squared kurtosis)}
\usage{
BoxCoxLambda(obs)
}
\arguments{
\item{obs}{The observations to be transformed}
}
\value{
The observations transformed

The estimated lambda
}
\description{
Performs Box-Cox transformation of the data. The transformation parameter (lambda) is chosen to minimize deviation from the normal distribution (minumum sum of squared skewness and squared kurtosis)
}
