% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{BoxCoxLambdaKnown}
\alias{BoxCoxLambdaKnown}
\title{Performs Box-Cox transformation with a known lambda parameter}
\usage{
BoxCoxLambdaKnown(obs, lambda)
}
\arguments{
\item{obs}{The observations to be transformed}

\item{lambda}{The lambda parameter to be used in the transformation}
}
\value{
The observations transformed
}
\description{
Performs Box-Cox transformation with a known lambda parameter
}
