% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{crpsEst}
\alias{crpsEst}
\title{Computes the Continuous Ranked Probability Score (CRPS)}
\usage{
crpsEst(obs, mean, sd, lambda, Nsamples = 10000, print2screen = TRUE)
}
\arguments{
\item{mean}{The mean of the fitted distribution}

\item{sd}{The standard deviation of the fitted distribution}

\item{lambda}{The lambda used in the Box Cox transformation}

\item{print2screen}{Logical parameter to turn on and off print to screen}

\item{obs.bc}{Box Cox transformed observations}
}
\value{
The Continuous Ranked Probability Score (CRPS)
}
\description{
Computes the Continuous Ranked Probability Score (CRPS)
}
