% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{dboxcox}
\alias{dboxcox}
\title{Compute density of the Box Cox distribution}
\usage{
dboxcox(obs, mean, sd, lambda, log = FALSE)
}
\arguments{
\item{obs}{The probability}

\item{mean}{The mean of the distribution}

\item{sd}{The standard deviation of the distribution}

\item{lambda}{The lambda used in the Box Cox transformation}

\item{log}{Logical value to switch between log or not}
}
\value{
The density values of the Box Cox distribution
}
\description{
Compute density of the Box Cox distribution
}
