% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fourier}
\alias{fourier}
\title{Calculates the Fourier terms for modeling seasonality}
\usage{
fourier(x, K = 2, m = 365.25 * 4)
}
\arguments{
\item{x}{Coefficients}

\item{K}{Number of Fourier terms (default = 2)}

\item{m}{Number of observations per period (default = 365.25*4)}
}
\value{
A matrix with 2xK Fourier terms
}
\description{
Calculates the Fourier terms for modeling seasonality
}
