% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{getPreddistr}
\alias{getPreddistr}
\title{Training the model and generating the predictive distributions}
\usage{
getPreddistr(
  SWH = NA,
  SLP = NA,
  SLP.grad = NA,
  latCell = 4,
  longCell = 4,
  neig = 2,
  na.thresh = 500,
  latSWH = NA,
  lonSWH = NA,
  latSLP = NA,
  longSLP = NA,
  intercept.fourier = NA,
  maxlag = 10
)
}
\arguments{
\item{SWH}{The SWH data}

\item{SLP}{The SLP data}

\item{SLP.grad}{The SLP gradient data}

\item{latCell}{Define the latitude cell of interest (default = 4)}

\item{longCell}{Define the longitude cell of interest (default = 4)}

\item{neig}{Define the number of spatial neighbours (default = 2)}

\item{na.thresh}{Define how many missing values is "good enough" (default = 500)}

\item{latSWH}{A vector containing the SWH latitudes}

\item{lonSWH}{A vector containing the SWH longitudes}

\item{latSLP}{A vector containing the SLP latitudes}

\item{intercept.fourier}{Seasonal Fourier coefficients}

\item{maxlag}{Define the maximum order of AR processes to be constructed}

\item{training.test}{The training and the test data}

\item{lonSLP}{A vector containing the SLO longitudes}
}
\value{
A list of predictive means, predictive spread and lambda for Box-Cox transformation
}
\description{
Training the model and generating the predictive distributions
}
