% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{graphDev}
\alias{graphDev}
\title{Prephare graphical device for a given OS}
\usage{
graphDev(width = 7, height = 5)
}
\arguments{
\item{width}{Width of the graphical window}

\item{height}{Height of the grapchical window}
}
\value{
The an OS dependent graphical window
}
\description{
Prephare graphical device for a given OS
}
