% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{InvBoxCox}
\alias{InvBoxCox}
\title{Compute the Inverse Box Cox tranformation}
\usage{
InvBoxCox(obst, lambda)
}
\arguments{
\item{obst}{Values to be detransformed}

\item{lambda}{The lambda used in the Box Cox transformation}
}
\value{
The quantiles of the Box Cox distribution
}
\description{
Compute the Inverse Box Cox tranformation
}
