% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{logsEst}
\alias{logsEst}
\title{Compute the log score}
\usage{
logsEst(obs, mean, sd, lambda, log = TRUE, print2screen = TRUE)
}
\arguments{
\item{obs}{Observations}

\item{mean}{The mean of the fitted distribution}

\item{sd}{The standard deviation of the fitted distribution}

\item{lambda}{The lambda used in the Box Cox transformation}

\item{print2screen}{Logical parameter to print results to screen}
}
\value{
The -log score. The lower value, the better.
}
\description{
Compute the log score
}
