% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{maeEst}
\alias{maeEst}
\title{Compute the mean absolute error}
\usage{
maeEst(obs, mean, sd, lambda, print2screen = TRUE)
}
\arguments{
\item{obs}{Observations}

\item{mean}{The mean of the fitted distribution}

\item{sd}{The standard deviation of the fitted distribution}

\item{lambda}{The lambda used in the Box Cox transformation}

\item{print2screen}{Logical parameter to print results to screen}
}
\value{
The mean absolute error of the Box Cox distribution
}
\description{
Compute the mean absolute error
}
