% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{makeARterms}
\alias{makeARterms}
\title{Calculates the AR lags}
\usage{
makeARterms(
  SWH.bc.standard.training,
  SWH.bc.fourier.training,
  SWH.bc.standard.test,
  SWH.bc.fourier.test,
  n.training = 100,
  n.test = 10,
  maxlag = 10
)
}
\arguments{
\item{n.training}{Size of training data}

\item{n.test}{Size of training data}

\item{maxlag}{Order of AR process}
}
\value{
AR terms for various lags used in model fitting
}
\description{
Calculates the AR lags
}
