% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{pBoxCox}
\alias{pBoxCox}
\title{Compute PIT values of the Box Cox distribution}
\usage{
pBoxCox(x, mean, sd, lambda, plothist = TRUE, graphD = FALSE)
}
\arguments{
\item{mean}{The mean of the predictive distribution}

\item{sd}{The standard deviation of the predictive distribution}

\item{lambda}{The estimated Box Cox lambda parameter used in the transformation of the predictive distribution}

\item{plothist}{Set to TRUE if plotting the histogram of the PIT values}

\item{obs}{The observations to be transformed}

\item{graphDev}{Logical variable to create a OS specific graphic device}
}
\value{
The PIT values
}
\description{
Compute PIT values of the Box Cox distribution
}
