% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{perfMeasures}
\alias{perfMeasures}
\title{Summary function to calculate a table of performance measures}
\usage{
perfMeasures(obs, mean, sd, lambda, Nsamples = 10000, print2screen = TRUE)
}
\arguments{
\item{obs}{Observations}

\item{mean}{The mean of the fitted distribution}

\item{sd}{The standard deviation of the fitted distribution}

\item{lambda}{The lambda for the Box-Cox distribution}

\item{Nsamples}{The number of samples to generate from the predictive distribution.}

\item{print2screen}{Logical parameter to print results to screen}
}
\value{
The root mean squared error of the predictive distribution
}
\description{
Summary function to calculate a table of performance measures
}
