% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{qBoxCox}
\alias{qBoxCox}
\title{Compute quantiles of the Box Cox distribution}
\usage{
qBoxCox(p, mean, sd, lambda)
}
\arguments{
\item{mean}{The mean of the distribution}

\item{sd}{The standard deviation of the distribution}

\item{lambda}{The lambda used in the Box Cox transformation}

\item{obs}{The probability}
}
\value{
The quantiles of the Box Cox distribution
}
\description{
Compute quantiles of the Box Cox distribution
}
