% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{rCorr}
\alias{rCorr}
\title{Plot random correlation for a selection of time points}
\usage{
rCorr(
  obs,
  t.period,
  mean,
  sd,
  lambda,
  n.random,
  training.test,
  SWHobs,
  graphD = FALSE
)
}
\arguments{
\item{obs}{The observations to be transformed}

\item{t.period}{The time period for plotting}

\item{mean}{The mean value of the distribution}

\item{sd}{The sd value of the distribution}

\item{lambda}{The estimated lambda of the distribution}

\item{n.random}{The number of random predictive trajectories}

\item{training.test}{The training and the test period}

\item{SWHobs}{SWH observations for a particular location in the test period}

\item{graphDev}{Logical variable to create a OS specific graphic device}
}
\value{
A figure showing the random correlations
}
\description{
Plot random correlation for a selection of time points
}
