% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{ribxEst}
\alias{ribxEst}
\title{Computes the reliability index}
\usage{
ribxEst(obs, mean, sd, lambda, n.bins = 20, print2screen = TRUE)
}
\arguments{
\item{mean}{The mean of the fitted distribution}

\item{sd}{The standard deviation of the fitted distribution}

\item{lambda}{The lambda used in the Box Cox transformation}

\item{n.bins}{Number of bins used}

\item{print2screen}{Logical parameter to print results to screen}

\item{obs.bc}{Box Cox transformed observations}
}
\value{
The reliability index (RIDX)
}
\description{
Computes the reliability index
}
