% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{rmseEst}
\alias{rmseEst}
\title{Compute the root mean squared error}
\usage{
rmseEst(obs, mean, sd, lambda, Nsamples = 10000, print2screen = TRUE)
}
\arguments{
\item{obs}{Observations}

\item{mean}{The mean of the fitted distribution}

\item{sd}{The standard deviation of the fitted distribution}

\item{lambda}{The lambda for the Box-Cox distribution}

\item{Nsamples}{The number of samples to generate from the predictive distribution.}

\item{print2screen}{Logical parameter to turn on and off print to screen}
}
\value{
The root mean squared error of the predictive distribution
}
\description{
Compute the root mean squared error
}
