% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{rPlotPred}
\alias{rPlotPred}
\title{Plot random predictive trajectories for a selection of time points}
\usage{
rPlotPred(obs, t.period, mean, sd, lambda, n.random, graphD = FALSE)
}
\arguments{
\item{obs}{The observations to be transformed}

\item{t.period}{The time period for plotting}

\item{mean}{The mean value of the distribution}

\item{sd}{The sd value of the distribution}

\item{lambda}{The estimated lambda of the distribution}

\item{n.random}{The number of random predictive trajectories}

\item{graphDev}{Logical variable to create a OS specific graphic device}
}
\value{
A figure showing the random predictive trajectories
}
\description{
Plot random predictive trajectories for a selection of time points
}
