% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{rpred}
\alias{rpred}
\title{Generate samples from the predictive distribution}
\usage{
rpred(n, mean, sd, lambda)
}
\arguments{
\item{mean}{Mean of untruncated normal distribution}

\item{sd}{Standard deviation of untruncated normal distribution}

\item{lambda}{paramter in Box Cox transformation}
}
\value{
Samples from the predictive distribution
}
\description{
Generate samples from the predictive distribution
}
