% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{rtnorm}
\alias{rtnorm}
\title{Compute the Inverse Box Cox tranformation}
\usage{
rtnorm(n, mean, sd, a, b)
}
\arguments{
\item{n}{Number of samples to simulate}

\item{mean}{Mean of truncated normal distribution}

\item{sd}{Standard deviation of truncated normal distribution}

\item{a}{Lower truncation limit}

\item{b}{Upper truncation limit}
}
\value{
Samples from the truncated normal distribution
}
\description{
Compute the Inverse Box Cox tranformation
}
