% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{split.data}
\alias{split.data}
\title{Splitting the data in a traing set and a test set}
\usage{
\method{split}{data}(years, trainingPeriod = 2006:2014, testPeriod = 2015)
}
\arguments{
\item{years}{A vector containg the year of each observation}

\item{trainingPeriod}{A vector defining for which time period the model is trained on (default = 2006:2014)}

\item{testPeriod}{Defining for which time period the model is tested on (default = 2015)}
}
\value{
A list containing the training set and the test set
}
\description{
Splitting the data in a traing set and a test set
}
